warehouse Inventory added and updated all components

+++ src/controllers/inventoryTransactionController.ts
@@ -3,26 +3,33 @@ import InventoryTransaction, {
   IInventoryTransaction,
 } from "../models/inventoryTransactionModel";
 import { Request, Response } from "express";
-import Product from "../models/productModel";
+import Product, { IProduct } from "../models/productModel";
 import User from "../models/userModel";
 import Supplier from "../models/supplierModel";
 import Order from "../models/orderModel";
+import Warehouse, { IWarehouse } from "../models/warehouseModel";
 
 //GET ALL INVENTORY TRANSACTIONS
 const getInventoryTransactions = async (
   req: Request,
   res: Response
-): Promise<void> => {
+): Promise<Response | undefined> => {
   try {
     const transactions: IInventoryTransaction[] =
-      await InventoryTransaction.find()
-        .populate("products", "name")
+      await InventoryTransaction.find({})
         .populate("staffId", "username")
-        .populate("supplierId", "name");
+        .populate("adminId", "username")
+        .populate("products", "name category price")
+        .populate("warehouseId", "name location capacity")
+        .populate("supplierId", "name contactInfo")
+        .populate("customerId", "username");
     console.log("Fetched transactions...");
-    res.status(200).json(transactions);
-  } catch (error) {
-    res.status(500).json({ error: "Could not fetched transactions" });
+    return res.status(200).json(transactions);
+  } catch (error: any) {
+    return res.status(500).json({
+      error: "Could not fetched transactions",
+      details: error.message,
+    });
   }
 };
 
@@ -35,18 +42,28 @@ const getInventoryTransaction = async (
     return res.status(404).json({ error: "Not a valid Document" });
   }
 
+  //Ensure that the Inventory exists
+  const inventoryExists = await InventoryTransaction.findById(req.params.id);
+  if (!inventoryExists) {
+    return res.status(404).json({ error: "Inventory not found" });
+  }
+
   try {
     const transaction: IInventoryTransaction | null =
       await InventoryTransaction.findById(req.params.id)
-        .populate("products", "name")
         .populate("staffId", "username")
-        .populate("supplierId", "name");
+        .populate("adminId", "username")
+        .populate("products", "name category price")
+        .populate("warehouseId", "name location capacity")
+        .populate("supplierId", "name contactInfo")
+        .populate("customerId", "username");
+    console.log("Fetched transactions...");
     if (!transaction) {
       return res.status(400).json({ error: "Document not found" });
     }
-    res.status(200).json(transaction);
+    return res.status(200).json(transaction);
   } catch (error) {
-    res.status(500).json({ error: "Could not fetch transaction" });
+    return res.status(500).json({ error: "Could not fetch transaction" });
   }
 };
 
@@ -55,19 +72,62 @@ const createInventoryTransaction = async (
   req: Request,
   res: Response
 ): Promise<Response | undefined> => {
-  const { transactionType, products, quantity, customerId, supplierId } =
-    req.body;
-  console.log(req.body);
+  const {
+    action,
+    transactionType,
+    fromWarehouseId,
+    toWarehouseId,
+    warehouseId,
+    products,
+    quantity,
+    customerId,
+    supplierId,
+    interWarehouseTransferStatus,
+    totalValue,
+  } = req.body;
+
   const staffId = (req as any).user.id;
+  const adminId = (req as any).user.id;
+
+  //Initialize Session Transaction
+  const session = await mongoose.startSession();
+  session.startTransaction();
+
+  //Check Input fields
   if (
     !transactionType ||
+    (["Inter-Warehouse Transfer", "Failed Transfer Request"].includes(
+      transactionType
+    ) &&
+      !fromWarehouseId) ||
+    (["Inter-Warehouse Transfer", "Failed Transfer Request"].includes(
+      transactionType
+    ) &&
+      !toWarehouseId) ||
     !products ||
     products.length === 0 ||
-    !quantity ||
-    (["Sales Transaction", "Customer Return"].includes(transactionType) &&
+    (["Inter-Warehouse Transfer", "Failed Transfer Request"].includes(
+      transactionType
+    ) &&
+      !staffId) ||
+    (["Sales Transaction", "Online Order", "Customer Return"].includes(
+      transactionType
+    ) &&
       !customerId) ||
     (["Restock Transaction", "Supplier Return"].includes(transactionType) &&
-      !supplierId)
+      !supplierId) ||
+    !transactionType ||
+    (["Addition/Removal of Product From Warehouse"].includes(transactionType) &&
+      !action) ||
+    ([
+      "Addition/Removal of Product From Warehouse",
+      "Restock Transaction",
+      "Supplier Return",
+      "Sales Transaction",
+      "Online Order",
+      "Customer Return",
+    ].includes(transactionType) &&
+      !warehouseId)
   ) {
     return res.status(400).json({ error: "Please fill all fields" });
   }
@@ -77,21 +137,45 @@ const createInventoryTransaction = async (
     const validTypes = [
       "Restock Transaction",
       "Sales Transaction",
+      "Online Order",
       "Damaged Product",
       "Supplier Return",
       "Customer Return",
+      "Inter-Warehouse Transfer",
+      "Addition/Removal of Product From Warehouse",
+      "Failed Transfer Request",
     ];
     if (!validTypes.includes(transactionType)) {
       return res.status(400).json({ error: "Invalid transaction type" });
     }
 
+    //Verify if Warehouse exists in database
+    const warehouse = await Warehouse.findById(warehouseId).session(session);
+    if (!warehouse) {
+      return res.status(404).json({ error: "Warehouse not found" });
+    }
+
+    //Check if the staff Id is the same as the warehouse managedBy
+    const isUserAuthorized = warehouse.managedBy.includes(staffId);
+    if (!isUserAuthorized) {
+      return res.status(403).json({
+        error: "User is not authorized to record this transaction",
+      });
+    }
+
     //Check if all product IDs are valid
-    const validProducts = await Product.find({ _id: { $in: products } });
+    const validProducts = await Product.find({
+      _id: {
+        $in: products.map(
+          (p: { productId: mongoose.Types.ObjectId }) => p.productId
+        ),
+      },
+    }).session(session);
     if (validProducts.length !== products.length) {
       return res.status(404).json({ error: "One or more products not found" });
     }
 
-    // Validate the product-supplier association
+    //Validate the product-supplier association
     for (let product of validProducts) {
       if (
         ["Restock Transaction", "Supplier Return"].includes(transactionType) &&
@@ -104,18 +188,65 @@ const createInventoryTransaction = async (
     }
 
     //Validate user existence
-    const user = await User.findById(staffId);
+    const user = await User.findById(staffId).session(session);
     if (!user) {
       return res.status(404).json({ error: "User not found" });
     }
 
-    //Validate customer ID for sales and customer returns
-    if (["Sales Transaction", "Customer Return"].includes(transactionType)) {
-      const customerOrder = await Order.findOne({
-        "products.productId": { $in: products },
-        user: customerId, // Ensure the customer ID matches the order
-      });
+    //Save new InventoryTransaction with product references
+    const productsArray = products.map(
+      (product: { productId: mongoose.Types.ObjectId; quantity: number }) => ({
+        productId: product.productId,
+        quantity: product.quantity,
+      })
+    );
 
+    //Check if the products are found in the warehouseId
+    const productsInWarehouse = await Product.find({
+      _id: {
+        $in: productsArray.map((product: any) => product.productId),
+      },
+      warehouse: warehouseId,
+    }).session(session);
+    if (productsInWarehouse.length !== productsArray.length) {
+      return res.status(400).json({
+        error: `One or more products do not belong to the specified warehouse.`,
+      });
+    }
+
+    //Validation for online order and customer returns
+    if (["Online Order", "Customer Return"].includes(transactionType)) {
+      //Check if the Order has been recorded in Inventory
+      const existingTransaction = await InventoryTransaction.findOne({
+        transactionType,
+        products,
+        customerId,
+      }).session(session);
+      if (existingTransaction) {
+        return res
+          .status(400)
+          .json({ error: "Order is already recorded in inventory" });
+      }
+
+      //Validate customer ID for Online Order and customer returns
+      const customerOrder = await Order.findOne({
+        //Check if each product matches both productId and quantity
+        $and: productsArray.map(
+          (product: {
+            productId: mongoose.Types.ObjectId;
+            quantity: number;
+          }) => ({
+            products: {
+              $elemMatch: {
+                productId: product.productId,
+                quantity: product.quantity,
+              },
+            },
+          })
+        ),
+        //Ensure the customer ID matches the order
+        user: customerId,
+      }).session(session);
       if (!customerOrder) {
         return res
           .status(404)
@@ -123,10 +254,36 @@ const createInventoryTransaction = async (
       }
     }
 
-    for (let product of validProducts) {
-      let updatedQuantity = product.quantity;
+    //Calculate totalValue and update product quantities based on transaction type
+    let totalValue = 0;
+    const productQuantities = [];
+    const warehouseUpdates = [];
+    //Now update totalValue and totalQuantity for each warehouse after updating product quantities
+    const warehouseProductUpdates: Record<
+      string,
+      { totalQuantity: number; totalValue: number }
+    > = {};
 
-      // HANDLE DIFFERENT TRANSACTION TYPES
+    //Update product quantities based on transaction type
+    for (let { productId, quantity } of productsArray) {
+      const product = validProducts.find((p) => p._id.equals(productId));
+      if (!product) {
+        return res.status(404).json({ error: "Product not found" });
+      }
+
+      let updatedQuantity = product.quantity;
+      let productPrice = product.price;
+      //Get the warehouse for this product
+      const warehouseId = product.warehouse.toString();
+
+      //Initialize warehouse update record if not already present
+      if (!warehouseProductUpdates[warehouseId]) {
+        warehouseProductUpdates[warehouseId] = {
+          totalQuantity: 0,
+          totalValue: 0,
+        };
+      }
+      //HANDLE DIFFERENT TRANSACTION TYPES
       switch (transactionType) {
         case "Restock Transaction":
           const restockSupplier = await Supplier.findById(supplierId);
@@ -134,14 +291,25 @@ const createInventoryTransaction = async (
             return res.status(404).json({ error: "Supplier not found" });
           }
           updatedQuantity += quantity;
+          totalValue += productPrice * quantity;
+          //Add to warehouse totalQuantity and totalValue
+          warehouseProductUpdates[warehouseId].totalQuantity += quantity;
+          warehouseProductUpdates[warehouseId].totalValue +=
+            productPrice * quantity;
           break;
 
         case "Sales Transaction":
+        case "Online Order":
         case "Damaged Product":
           if (product.quantity < quantity) {
             return res.status(400).json({ error: "Insufficient stock" });
           }
           updatedQuantity -= quantity;
+          totalValue += productPrice * quantity;
+          //Subtract from warehouse totalQuantity and totalValue
+          warehouseProductUpdates[warehouseId].totalQuantity -= quantity;
+          warehouseProductUpdates[warehouseId].totalValue -=
+            productPrice * quantity;
           break;
 
         case "Supplier Return":
@@ -153,29 +321,57 @@ const createInventoryTransaction = async (
             return res.status(400).json({ error: "Insufficient stock" });
           }
           updatedQuantity -= quantity;
+          totalValue += productPrice * quantity;
+          //Subtract from warehouse totalQuantity and totalValue
+          warehouseProductUpdates[warehouseId].totalQuantity -= quantity;
+          warehouseProductUpdates[warehouseId].totalValue -=
+            productPrice * quantity;
           break;
 
         case "Customer Return":
           updatedQuantity += quantity;
+          totalValue += productPrice * quantity;
+          //Add to warehouse totalQuantity and totalValue
+          warehouseProductUpdates[warehouseId].totalQuantity += quantity;
+          warehouseProductUpdates[warehouseId].totalValue +=
+            productPrice * quantity;
           break;
+
+        // case "Inter-Warehouse Transfer":
+        //   if (product.quantity < quantity) {
+        //     return res.status(400).json({ error: "Insufficient stock" });
+        //   }
+        //   updatedQuantity -= quantity;
+        //   break;
 
         default:
           return res.status(400).json({ error: "Invalid transaction type" });
       }
+      //Prepare product quantity update
+      productQuantities.push({
+        updateOne: {
+          filter: { _id: new mongoose.Types.ObjectId(product._id) },
+          update: { $set: { quantity: updatedQuantity } },
+        },
+      });
 
-      //Update product quantity
-      product.quantity = updatedQuantity;
-      await product.save();
+      //Update product quantity in the Warehouse collection
+      warehouseUpdates.push({
+        updateOne: {
+          filter: { _id: warehouseId, "products.productId": productId },
+          update: {
+            $set: { "products.$.quantity": updatedQuantity },
+          },
+        },
+      });
     }
 
     //Save new Inventory transaction
     const newInventory: IInventoryTransaction = new InventoryTransaction({
       transactionType,
-      products,
+      products: productsArray,
       quantity,
-      customerId: ["Sales Transaction", "Customer Return"].includes(
-        transactionType
-      )
+      customerId: ["Online Order", "Customer Return"].includes(transactionType)
         ? customerId
         : undefined, //Optional based on transaction type
       staffId,
@@ -184,15 +380,75 @@ const createInventoryTransaction = async (
       )
         ? supplierId
         : undefined, //Optional based on transaction type
+      toWarehouseId: [
+        "Inter-Warehouse Transfer",
+        "Failed Transfer Request",
+      ].includes(transactionType)
+        ? toWarehouseId
+        : undefined, //Optional based on transaction type
+      fromWarehouseId: [
+        "Inter-Warehouse Transfer",
+        "Failed Transfer Request",
+      ].includes(transactionType)
+        ? fromWarehouseId
+        : undefined, //Optional based on transaction type
+      action: ["Addition/Removal of Product From Warehouse"].includes(
+        transactionType
+      )
+        ? action
+        : undefined, //Optional based on transaction type
+      warehouseId: [
+        "Addition/Removal of Product From Warehouse",
+        "Online Order",
+        "Customer Return",
+        "Restock Transaction",
+        "Supplier Return",
+        "Sales Transaction",
+      ].includes(transactionType)
+        ? warehouseId
+        : undefined, //Optional based on transaction type
+      interWarehouseTransferStatus: ["Inter-Warehouse Transfer"].includes(
+        transactionType
+      )
+        ? interWarehouseTransferStatus
+        : undefined, //Optional based on transaction type
+      totalValue,
     });
-    await newInventory.save();
-    console.log("Transaction created...");
-    res.status(201).json(newInventory);
+
+    // Save the new Inventory Update product quantities and Update Warehouse
+    await newInventory.save({ session });
+    await Product.bulkWrite(productQuantities, { session });
+    await Warehouse.bulkWrite(warehouseUpdates, { session });
+
+    //Update the warehouse totalQuantity and totalValue after processing all products
+    for (const [warehouseId, { totalQuantity, totalValue }] of Object.entries(
+      warehouseProductUpdates
+    )) {
+      await Warehouse.findByIdAndUpdate(
+        warehouseId,
+        {
+          $inc: {
+            totalQuantity,
+            totalValue,
+          },
+        },
+        { session }
+      );
+    }
+    //Commit the transaction
+    await session.commitTransaction();
+
+    return res.status(201).json(newInventory);
   } catch (error: any) {
-    res.status(500).json({
+    //If an error occurs, abort the transaction
+    await session.abortTransaction();
+    return res.status(500).json({
       error: "Failed to create new inventory transaction",
       details: error.message,
     });
+  } finally {
+    //End transaction
+    session.endSession();
   }
 };
 
@@ -201,18 +457,47 @@ const updateInventoryTransaction = async (
   req: Request,
   res: Response
 ): Promise<Response | undefined> => {
-  const { transactionType, productId, quantity, customerId, supplierId } =
-    req.body;
+  const {
+    action,
+    transactionType,
+    fromWarehouseId,
+    toWarehouseId,
+    warehouseId,
+    products,
+    quantity,
+    customerId,
+    supplierId,
+    interWarehouseTransferStatus,
+  } = req.body;
 
   if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
     return res.status(404).json({ error: "Not a valid document" });
+  }
+
+  //Check if Inventory Transaction exists
+  const inventoryTransactionExists = await InventoryTransaction.findById(
+    req.params.id
+  );
+  if (!inventoryTransactionExists) {
+    return res.status(404).json({ error: "Inventory transaction not found" });
   }
 
   try {
     const updatedInventory: IInventoryTransaction | null =
       await InventoryTransaction.findByIdAndUpdate(
         req.params.id,
-        { transactionType, productId, quantity, customerId, supplierId },
+        {
+          action,
+          transactionType,
+          fromWarehouseId,
+          toWarehouseId,
+          warehouseId,
+          products,
+          quantity,
+          customerId,
+          supplierId,
+          interWarehouseTransferStatus,
+        },
         { new: true }
       );
 
@@ -234,6 +519,14 @@ const deleteInventoryTransaction = async (
     return res.status(404).json({ error: "Not a valid document" });
   }
 
+  //Check if Inventory Transaction exists
+  const inventoryTransactionExists = await InventoryTransaction.findById(
+    req.params.id
+  );
+  if (!inventoryTransactionExists) {
+    return res.status(404).json({ error: "Inventory transaction not found" });
+  }
+
   try {
     const deletedInventoryTransaction: IInventoryTransaction | null =
       await InventoryTransaction.findByIdAndDelete(req.params.id);
@@ -243,8 +536,11 @@ const deleteInventoryTransaction = async (
     res
       .status(201)
       .json({ message: "Inventory transaction deleted successfully" });
-  } catch (error) {
-    res.status(500).json({ error: "Could not delete Inventory transaction" });
+  } catch (error: any) {
+    res.status(500).json({
+      error: "Could not delete Inventory transaction",
+      details: error.message,
+    });
   }
 };
 
@@ -255,8 +551,3 @@ export {
   updateInventoryTransaction,
   deleteInventoryTransaction,
 };
-
-// const product = await Product.findById(productId);
-// if (!product) {
-//   return res.status(404).json({ error: "Product not found" });
-// }
